コードレビュー後の修正
destroyアクション内にログインしているユーザーかどうかを識別するためのメソッドを追加。

+++ app/controllers/items_controller.rb
@@ -35,10 +35,13 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
-    else
-      render :show
+    binding.pry
+    if @item.user_id == current_user.id
+      if @item.destroy
+        redirect_to root_path
+      else
+        render :show
+      end
     end
   end
